backend/requirements.txt
pillow
zeroscratches  # or your custom scratch model
